Move Hardware file with PTAL for reusability
