+++ SimpleSampler/firmwares/SimpleSampler/debug.h
@@ -3,7 +3,6 @@
 
 
 #include "./constants.h"
-#include "./Hardware.h"
 
 
 //▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄
